backend/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.10.1
pymysql
sqlparse==0.5.3
psycopg2-binary
python-dotenv==1.0.1